change data type

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -6,11 +6,11 @@
  */
 int main(void)
 {
-	int i;
-	int first = 1;
-	int second = 2;
-	int next_num;
-	int sum;
+	long i;
+	long first = 1;
+	long second = 2;
+	long next_num;
+	long sum;
 
 	sum += second;
 	for (i = 3; i <= 50; i++)
@@ -26,7 +26,7 @@ int main(void)
 			second = next_num;
 		}
 	}
-	printf("%d", sum);
+	printf("%ld", sum);
 	putchar('\n');
 	return (0);
 }
